refactor : 테스트 코드 작성 및 Exception 위치 변경
