+++ src/main/java/lottogame/domain/InvalidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lottogame.lottogameexception;
+package lottogame.domain;
 
 public class InvalidLottoNumberException extends RuntimeException {
     public InvalidLottoNumberException(String message) {
