+++ src/main/java/lottogame/domain/InvalidLottoPriceException.java
@@ -1,4 +1,4 @@
-package lottogame.lottogameexception;
+package lottogame.domain;
 
 public class InvalidLottoPriceException extends RuntimeException {
     public InvalidLottoPriceException(String message) {
