+++ src/test/java/lottogame/utils/LottoNumbersParserTest.java
@@ -1,6 +1,6 @@
 package lottogame.utils;
 
-import lottogame.lottogameexception.InvalidLottoNumberException;
+import lottogame.domain.InvalidLottoNumberException;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
